# packages/mtmai/mtmai-0.3.1174.tar.gz/mtmai-0.3.1174/mtmai/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mtmai",
    "description": "mtmai description(group)",
    "version": "0.3.1174"
  },
  "paths": {
...
    "/api/v1/chat/config": {
      "post": {
        "tags": [
          "openai"
        ],
        "summary": "Chat Config",
        "operationId": "openai-chat_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssisantConfig"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/chat/completions": {
      "post": {
        "tags": [
          "openai"
        ],
        "summary": "Chat Completions",
        "description": "\u4ee5\u517c\u5bb9 openai completions \u534f\u8bae\u7684\u65b9\u5f0f \u53cd\u4ee3\u5176\u4ed6 ai \u63d0\u4f9b\u5546\u7684 completions api",
        "operationId": "openai-chat_completions",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssisantConfig": {
        "properties": {
          "chatProfile": {
            "type": "string",
            "title": "Chatprofile"
          },
          "starts": {
            "items": {
              "$ref": "#/components/schemas/AssisantStart"
            },
            "type": "array",
            "title": "Starts"
          },
          "workbenchs": {
            "items": {
              "$ref": "#/components/schemas/AssisantWorkBrench"
            },
            "type": "array",
            "title": "Workbenchs"
          },
          "workbench_default": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workbench Default"
          },
          "markdown_description_tpl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Markdown Description Tpl"
          }
        },
        "type": "object",
        "required": [
          "chatProfile",
          "starts",
          "workbenchs"
        ],
        "title": "AssisantConfig"
      },
      "AssisantStart": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "AssisantStart"
      },
      "AssisantWorkBrench": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "description"
        ],
        "title": "AssisantWorkBrench"
      },
...
      "ChatConfigRequest": {
        "properties": {
          "chatProfile": {
            "type": "string",
            "title": "Chatprofile"
          }
        },
        "type": "object",
        "required": [
          "chatProfile"
        ],
        "title": "ChatConfigRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  },
...
}